fix: use small integer for channel status

+++ api/src/subscription.rs
@@ -804,11 +804,11 @@ impl SubscriptionRoot {
         let source_i64 = source_key_id.map(|k| k as i64);
         let dest_i64 = destination_key_id.map(|k| k as i64);
 
-        // Convert GraphQL ChannelStatus to internal status code (i8)
+        // Convert GraphQL ChannelStatus to internal status code (i16)
         let status_code = status.map(|s| match s {
-            blokli_api_types::ChannelStatus::Closed => 0i8,
-            blokli_api_types::ChannelStatus::Open => 1i8,
-            blokli_api_types::ChannelStatus::PendingToClose => 2i8,
+            blokli_api_types::ChannelStatus::Closed => 0i16,
+            blokli_api_types::ChannelStatus::Open => 1i16,
+            blokli_api_types::ChannelStatus::PendingToClose => 2i16,
         });
 
         // Use optimized channel aggregation function
@@ -933,7 +933,7 @@ mod tests {
         tx_index: i64,
         log_index: i64,
         balance: Vec<u8>,
-        status: i8,
+        status: i16,
     ) -> anyhow::Result<i64> {
         let state = blokli_db_entity::channel_state::ActiveModel {
             id: Default::default(),
